JS runs in the browser
previously it only ran in `jest` unit tests

- removed `out`; we weren't using it. It was cruft.
- removed extraneous `console.log(...`; it was unneeded.
- guarded against `module` not being defined.

fixes:
```
ReferenceError: module is not defined
```

To make JS work in the browser, we didn't need to
`require(charlotte.js);` even though the template for jest insisted we
have it. They lie! I can't believe they lied to us. We merely needed to
bring in the script from `index.html` (`<script src="...`).

Signed-off-by: Brian Cunnie <brian.cunnie@gmail.com>